Move enum classes into dedicated file
